# api/frontend/public/api_doc.json
{
  "swagger": "2.0",
  "info": {
    "description": "API Documentation for MINT Data Catalog",
    "version": "1.0.0",
    "title": "MINT Data Catalog",
    "contact": {
      "email": "danf@usc.edu"
    }
  },
  "host": "data-catalog.mint.isi.edu",
...
  "schemes": ["https"],
  "paths": {
...
    "/resources/get_resource_info": {
      "post": {
        "tags": ["Resources"],
        "summary": "Detailed information about the resource",
        "description": "",
        "operationId": "getResourceInfo",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "dataset_id": {
                  "type": "string",
                  "format": "uuid",
                  "description": "record_id of the resource"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "dataset": {
                  "type": "object",
                  "properties": {
                    "resource_id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "name": {
                      "type": "string"
                    },
                    "dataset_id": {
                      "type": "string",
                      "format": "uuid",
                      "description": "record_id of the dataset that this resource belongs to"
                    },
                    "resource_type": {
                      "type": "string"
                    },
                    "data_url": {
                      "type": "string",
                      "format": "uri"
                    },
                    "metadata": {
                      "type": "object"
                    },
                    "created_at": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input (check error message(s) to correct it)"
          },
          "500": {
            "description": "Internal error - please contact Dan Feldman danf@usc.edu"
          }
        }
      }
    },
    "/resources/update_resource": {
      "post": {
        "tags": ["Resources"],
        "summary": "Update resource",
        "description": "",
        "operationId": "updateResource",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Resource definition",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["resource_id"],
              "properties": {
                "resource_id": {
                  "type": "string",
                  "format": "uuid",
                  "description": "Unique identifier"
                },
                "name": {
                  "type": "string",
                  "description": "New display name/title of the resource"
                },
                "resource_type": {
                  "type": "string",
                  "enum": [
                    "GeoJSON",
                    "NetCDF",
                    "CSV",
                    "GeoTIFF",
                    "JSON",
                    "SHP",
                    "TXT",
                    "<Write-in your own>"
                  ]
                },
                "data_url": {
                  "type": "string",
                  "description": "New direct download url for this resource"
                },
                "metadata": {
                  "type": "object"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Invalid input (check error message(s) to correct it)"
          },
          "500": {
            "description": "Internal error - please contact Dan Feldman danf@usc.edu"
          }
        }
      }
    },
...
  },
...
}